Add header component to queryeditor
