+++ src/VariableQueryEditor.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { RedshiftQuery, RedshiftDataSourceOptions } from './types';
 import { QueryEditorProps } from '@grafana/data';
 import { DataSource } from 'datasource';
-import { QueryEditor } from 'QueryEditor';
+import { QueryEditorForm } from 'QueryEditorForm';
 
 export function VariableQueryCodeEditor(props: QueryEditorProps<DataSource, RedshiftQuery, RedshiftDataSourceOptions>) {
-  return <QueryEditor {...props} hideRunQueryButtons></QueryEditor>;
+  return <QueryEditorForm {...props}></QueryEditorForm>;
 }
